Added Test case and renamed main.py
