+++ README.md
@@ -19,7 +19,7 @@ You only need to duplicate the tabs and keep the format for each of the people y
 1. Create a folder and place the attached script & Users.xlsx file in the folder (download this Google Sheet here and rename Users.xlsx) - Or change line 48 if you don't want to rename.
 2. Open a terminal and cd to that folder
 3. And then run the below command
-```python main.py```
-4. This will generate an Assesments.html file inside the same directory, all of the diagrams will be in a single html page labelled, scroll down.
+```python PolarAreaChart.py```
+4. This will generate an Assessments.html file inside the same directory, all of the diagrams will be in a single html page labelled, scroll down.
 You only need to duplicate the tabs and keep the format for each of the people you want to perform an assessment of. Label each tab with the person's name as the script uses this as a label for the diagrams.
 5. To edit, change column B in Users.xlsx. Use a new tab for each person.
